Add files via upload due to Problems with my Visual studio on Mac
Sorry Professor, 
My Visual Studio has been giving me issues and even my firs commit had to be forced and pushed unnecessary files, which is why my first commit message is with a different old project (NewLogin) that was pushed along with the Applied activity. I don't have time to start recreating another repo and recommitting so I ended up zipping and uploading later on at the end. 

COMMIT MESSAGE:
On the student and course pages, save and add buttons have been added. Additionally, when all courses or students are dropped, the home page will now display no courses or students. This is because, rather than having the course or student removed from the list used for display, the LoadData function is called when a course or student is deleted. The student model has been updated to use a courseId rather than a classid. The repository now includes update methods for both courses and students. When there are no courses in the database, it will now be indicated in the picker on the student page. Comments were added. discarded unnecessary student courses classes.